@@ -4,7 +4,7 @@ class InfoText extends StatelessWidget {
   final String type;
   final String text;
 
-  InfoText({@required this.type, @required this.text});
+  InfoText({@required required this.type, required this.text});
 
   @override
   Widget build(BuildContext context) {
