@@ -5,16 +5,14 @@ import 'package:jordyhers/utils/theme.dart';
 class CenteredView extends StatelessWidget {
   final Widget child;
 
-  const CenteredView({Key key, @required this.child}) : super(key: key);
+  const CenteredView({Key? key, @required required this.child})
+      : super(key: key);
 
   @override
   Widget build(BuildContext context) {
     var screenSize = MediaQuery.of(context).size;
     return Container(
-      decoration: BoxDecoration(
-        image: DecorationImage(
-            image: NetworkImage(st.backgroundImage), fit: BoxFit.fitWidth),
-      ),
+      color: Colors.black,
       alignment: Alignment.topCenter,
       child: ConstrainedBox(
         constraints: BoxConstraints(
