@@ -5,7 +5,7 @@ import 'navigation_bar_mobile.dart';
 import 'navigation_bar_tablet_desktop.dart';
 
 class NavigationBar extends StatelessWidget {
-  const NavigationBar({Key key}) : super(key: key);
+  const NavigationBar({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
