@@ -1,5 +1,7 @@
 import 'package:flutter/material.dart';
+import 'package:jordyhers/services/url_launcher.dart';
 import 'package:jordyhers/view/layout_template.dart';
+import 'package:provider/provider.dart';
 import 'locator.dart';
 
 void main() {
@@ -16,7 +18,7 @@ class MyApp extends StatelessWidget {
       debugShowCheckedModeBanner: false,
       title: 'Jordy Hers',
       theme: ThemeData(primarySwatch: Colors.indigo, fontFamily: 'Montserrat'),
-      home: LayoutTemplate(),
+      home: Provider(create: (_) => Http(), child: LayoutTemplate()),
     );
   }
 }
