@@ -17,7 +17,7 @@ class Style {
   ).createShader(Rect.fromLTWH(400.0, 10.0, 300.0, 14.0));
 
   static final LinearGradient gradient = LinearGradient(colors: [
-    Colors.purpleAccent,
     Colors.blueAccent,
-  ]);
+    Colors.purpleAccent,
+  ], begin: Alignment.topLeft, end: Alignment.bottomRight);
 }
